Move helper methods related to the product table into a separate class.
