--- src/main/java/ro/sellfluence/db/ProductInfo.java
@@ -1,13 +1,0 @@
-package ro.sellfluence.db;
-
-public record ProductInfo(
-        String pnk,
-        String productCode,
-        String name,
-        boolean continueToSell,
-        boolean retracted,
-        String category,
-        String messageKeyword,
-        String employeeSheetName
-) {
-}
